YouTube Downloader/workspace/requirements.txt
yt-dlp>=2024.1.1
customtkinter>=5.2.2
Pillow>=10.2.0
requests>=2.31.0
ffmpeg-python>=0.2.0
